+++ tests/scheduler/backends/test_qblox_backend.py
@@ -641,6 +641,11 @@ def test_find_all_port_clock_combinations(load_example_qblox_hardware_config):
         ("q5:mw", "q5.01"),
         ("q4:res", "q4.ro"),
         ("q5:res", "q5.ro"),
+        ("q0:fl", "cl0.baseband"),
+        ("q1:fl", "cl0.baseband"),
+        ("q2:fl", "cl0.baseband"),
+        ("q3:fl", "cl0.baseband"),
+        ("q4:fl", "cl0.baseband"),
     }
     assert portclocks == answer
 
@@ -650,7 +655,7 @@ def test_generate_port_clock_to_device_map(load_example_qblox_hardware_config):
         load_example_qblox_hardware_config
     )
     assert (None, None) not in portclock_map.keys()
-    assert len(portclock_map.keys()) == 13
+    assert len(portclock_map.keys()) == 18
 
 
 # --------- Test classes and member methods ---------